Smoothen animated widgets
