+++ src/main/java/gregtech/api/gui/widgets/ProgressWidget.java
@@ -1,9 +1,9 @@
 package gregtech.api.gui.widgets;
 
 import gregtech.api.gui.IRenderContext;
+import gregtech.api.gui.TimedProgressSupplier;
 import gregtech.api.gui.Widget;
 import gregtech.api.gui.resources.TextureArea;
-import gregtech.api.util.GTLog;
 import gregtech.api.util.Position;
 import gregtech.api.util.Size;
 import net.minecraft.network.PacketBuffer;
@@ -151,29 +151,4 @@ public class ProgressWidget extends Widget {
         }
     }
 
-    private static class TimedProgressSupplier implements DoubleSupplier {
-
-        private final int msPerCycle;
-        private final int maxValue;
-        private final boolean countDown;
-        private final long startTime;
-
-        public TimedProgressSupplier(int ticksPerCycle, int maxValue, boolean countDown) {
-            this.msPerCycle = ticksPerCycle * 50;
-            this.maxValue = maxValue;
-            this.countDown = countDown;
-            this.startTime = System.currentTimeMillis();
-        }
-
-        @Override
-        public double getAsDouble() {
-            long currentTime = System.currentTimeMillis();
-            long msPassed = (currentTime - startTime) % msPerCycle;
-            int currentValue = (int) Math.floorDiv(msPassed * (maxValue + 1), msPerCycle);
-            if (countDown) {
-                return (maxValue - currentValue) / (maxValue * 1.0);
-            }
-            return currentValue / (maxValue * 1.0);
-        }
-    }
 }
