+++ src/main/java/gregtech/api/terminal/gui/widgets/AnimaWidgetGroup.java
@@ -26,11 +26,15 @@ public abstract class AnimaWidgetGroup extends WidgetGroup {
     }
 
     @Override
-    public void updateScreen() {
+    public void updateScreenOnFrame() {
         if (interpolator != null) {
             interpolator.update();
         }
         super.updateScreen();
+    }
+
+    @Override
+    public void updateScreen() {
     }
 
     @SideOnly(Side.CLIENT)
